requirements.txt
# requirements.txt - Python dependencies for Crop Disease Prediction System

# Core Flask framework and extensions
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
Flask-Limiter==3.5.0
Flask-Babel==4.0.0

# Monitoring and Observability
prometheus-flask-exporter==0.22.4
prometheus-client==0.17.1

# Authentication and Security
PyJWT==2.8.0

# ML and Image Processing
tensorflow==2.15.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
opencv-python==4.8.0.76
numpy==1.24.3
scikit-image==0.21.0

# LLM Integration
google-generativeai==0.2.2
ollama==0.1.7
requests==2.31.0

# Caching and Sessions
redis==4.6.0

# Background Task Processing
celery==5.3.1

# System Monitoring
psutil==5.9.5

# Configuration and Environment
python-dotenv==1.0.0

# Data Processing and Analysis
pandas==2.0.3
matplotlib==3.7.2

# Development and Testing
pytest==7.4.0
pytest-flask==1.2.0

# ML Training and Optimization (for training notebook)
optuna==3.5.0
tensorflow-model-optimization==0.8.0
pytest-mock==3.11.1
Babel==2.12.1

# Logging and Structured Logging
structlog==23.1.0

# Model optimization and deployment
tf2onnx==1.14.0
tflite-runtime==2.13.0

# Web Server (for production deployment)
gunicorn==21.2.0
gevent==23.7.0

# Additional utilities
werkzeug==2.3.7
click==8.1.7

# Optional: For advanced image processing
albumentations==1.3.1

# Optional: For model quantization
tensorflow-model-optimization==0.7.5